Refactored article urls to slugs insted of ids
